refactor: update, insert 클래스 추출
